@MuriloAlmeid subindo post do usuario parte cuidador

+++ src/main/java/sptech/elderly/repository/UsuarioRepository.java
@@ -3,5 +3,5 @@ package sptech.elderly.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sptech.elderly.entity.UsuarioEntity;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
+public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Integer> {
 }
